[module7-task1]: Move comments-related E2E tests to comments.router
