--- src/service/server/routes/.eslintrc.yml
@@ -1,12 +1,0 @@
-extends:
-  - "../../../../.eslintrc.yml"
-parserOptions:
-  project: "../../tsconfig.json"
-overrides:
-  - files:
-      - "*.e2e.test.ts"
-    rules:
-      "@typescript-eslint/no-unsafe-member-access":
-        - off
-      "@typescript-eslint/no-unsafe-call":
-        - off
